Gambiarra

+++ src/contas/Conta.java
@@ -30,7 +30,7 @@ public abstract class  Conta {
 	public void setSaldo(double saldo) {
 		this.saldo = saldo;
 	}
-	public double getNumero() {
+	public long getNumero() {
 		return numero;
 	}
 	public void setNumero(long numero) {
